Move `apt_sources` and `yum_sources` tables to linux only (#7537)
