+++ specs/CMakeLists.txt
@@ -174,13 +174,14 @@ function(generateNativeTables)
     "linux/apparmor_profiles.table:linux"
     "linux/apparmor_events.table:linux"
     "linux/systemd_units.table:linux"
+    "linux/apt_sources.table:linux"
+    "linux/yum_sources.table:linux"
     "linwin/intel_me_info.table:linux,windows"
     "linwin/secureboot.table:linux,windows"
     "kernel_info.table:linux,macos,windows"
     "macwin/certificates.table:macos,windows"
     "posix/extended_attributes.table:linux,macos"
     "posix/acpi_tables.table:linux,macos"
-    "posix/apt_sources.table:linux,macos,freebsd"
     "posix/augeas.table:linux,macos,freebsd"
     "posix/authorized_keys.table:linux,macos,freebsd"
     "posix/block_devices.table:linux,macos"
@@ -245,7 +246,6 @@ function(generateNativeTables)
     "posix/ulimit_info.table:linux,macos,freebsd"
     "posix/usb_devices.table:linux,macos"
     "posix/user_events.table:linux,macos,freebsd"
-    "posix/yum_sources.table:linux,macos,freebsd"
     "sleuthkit/device_file.table:linux,macos,freebsd,windows"
     "sleuthkit/device_hash.table:linux,macos,freebsd,windows"
     "sleuthkit/device_partitions.table:linux,macos,freebsd,windows"
